+++ sns.go
@@ -1,20 +1,33 @@
-package twitter
+package main
 
 import (
+	"bytes"
+	"encoding/json"
 	"fmt"
 	"io/ioutil"
+	"net/http"
 	"os"
 	"strings"
 
 	"github.com/dghubble/oauth1"
 )
 
+type Misskey struct {
+	token string
+}
+
+type ReqBody struct {
+	I      string `json:"i"`
+	Text   string `json:"text"`
+	Detail bool   `json:"detail"`
+}
+
 type Twitter struct {
 	vid   string
 	title string
 }
 
-func New() *Twitter {
+func NewTwitter() *Twitter {
 	return &Twitter{}
 }
 
@@ -46,5 +59,36 @@ func (tw *Twitter) Tweet() error {
 	defer resp.Body.Close()
 	body, _ := ioutil.ReadAll(resp.Body)
 	fmt.Printf("Raw Response Body:\n%v\n", string(body))
+	return nil
+}
+
+func NewMisskey(token string) *Misskey {
+	return &Misskey{token: token}
+}
+
+func (m *Misskey) Post(id string, title string) error {
+	url := "https://@aopontan@misskey.io/api/notes/create"
+	content := fmt.Sprintf(`
+	【5分後に公開】
+	%s
+	https://www.youtube.com/watch?v=%s
+	`, title, id)
+
+	resb := ReqBody{
+		I:      m.token,
+		Text:   content,
+		Detail: false,
+	}
+
+	payload, err := json.Marshal(resb)
+	if err != nil {
+		return err
+	}
+
+	resp, err := http.Post(url, "application/json", bytes.NewBuffer(payload))
+	if err != nil {
+		return err
+	}
+	defer resp.Body.Close()
 	return nil
 }
